Modify SaveNativeEventToHALQueue() fix
Allow calls to SaveNativeEventToHALQueue() from ISR in the case that a
continuation is being run. This should be the most common case by far.
We assume this is safe because continuations rarely muck with managed
memory and when they do, they GLOBAL_LOCK() first... I hope...

Examples of touching managed memory are SaveNativeEventToHALQueue() and
private_malloc(). The latter works poorly and its use is discouraged in
general...