USB: Always return success even on drop
Theory is that the CLR is upset by the Debug.Prints() not indicating
success and is causing instability if this goes on for too long.

This can happen in particular if the host USB serial port is connected
(plugged-in) but does not serivce the port.

Workaround is to never admit failure. This is not a real issue with USB
anyway.